:rotating_light: Fix Lint for error
